Rename package testbed to slavsquatsuperstar.test
Also added documentation for Texture(GL) classes
