+++ src/test/java/slavsquatsuperstar/test/MainTest.java
@@ -1,4 +1,4 @@
-package testbed;
+package slavsquatsuperstar.test;
 
 import java.util.ArrayList;
 import java.util.List;
