+++ src/test/java/slavsquatsuperstar/test/collisiontests/OBBTests.java
@@ -1,4 +1,4 @@
-package testbed.collisiontests;
+package slavsquatsuperstar.test.collisiontests;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
