+++ src/test/java/slavsquatsuperstar/test/collisiontests/PolygonTests.java
@@ -1,4 +1,4 @@
-package testbed.collisiontests;
+package slavsquatsuperstar.test.collisiontests;
 
 import org.junit.jupiter.api.Test;
 import slavsquatsuperstar.math.Vec2;
