+++ src/test/java/slavsquatsuperstar/test/iotests/TextFileTests.java
@@ -1,4 +1,4 @@
-package testbed.iotests;
+package slavsquatsuperstar.test.iotests;
 
 import org.junit.jupiter.api.Test;
 import slavsquatsuperstar.fileio.AssetType;
