+++ src/test/resources/testassets/foobar/foobar.json
@@ -1,4 +1,4 @@
 {
     "date": "2021-09-12",
-    "time": "17:15:36.015652"
+    "time": "19:46:29.290250"
 }
