+++ src/test/resources/testassets/foobar/foobar.txt
@@ -1,2 +1,2 @@
 date=2021-09-12
-time=17:15:35.999333
+time=19:46:29.279490
